Allow cmdbuffer creation from compositor thread.
This adds a message filter on the IO thread which routes messages to
the corresponding cmdbuffer proxy on the correct thread (main or compositor).
The proxies become reference counted so that the filter can track them.
When a context gets deleted, a flag gets set on the context to indicate
they have become orphaned so we do not handle any incoming messages anymore
(the callbacks etc. have become invalid when the context goes away).


Review URL: http://codereview.chromium.org/7634019

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@97824 0039d316-1c4b-4281-b951-d872f2087c98