touchui: Split a KeyboardWidget out of KeyboardManager.
Since the KeyboardManager is a Singleton, it gets destroyed after all of views
goes away. Trying to destroy it (as a Widget) at that point causes problems that
can lead to crashes on shutdown. So split out the Widget and destroy it when chrome
is terminating to avoid such crashes.

BUG=None
TEST=None

Review URL: http://codereview.chromium.org/7721004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@98052 0039d316-1c4b-4281-b951-d872f2087c98