Fix missing null check in Clipboard::ReadImage.
Even though CF_DIB is always supposed to be synthesized from CF_BITMAP, change
the types to match in Clipboard::ReadAvailableTypes. We also add a NULL check
in Clipboard::ReadImage in case the contents of the clipboard change between the
time the clipboard types are accessed and the the clipboard content is accessed.

BUG=94332
TEST=none

Review URL: http://codereview.chromium.org/7741035

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@98361 0039d316-1c4b-4281-b951-d872f2087c98