For MergeUrls(), convert anonymous enum that gives bit values into a typedefed uint32.  This allows places wishing to refer to "a bitfield composed of these values" to use an explicit type.  (This isn't possible by simply naming the enum as technically the enum doesn't define all of the possible combinations of bits.)  This also means the individual named bit constants themselves have the same explicit type.
The drawback to defining constants at their declaration point is that supplying them to a templated function, like what EXPECT_EQ() expands into, triggers an "undefined symbol" error on Mac/Linux (and adding explicit storage for them in the .cc file can cause duplicate symbol errors on Windows).  Here I've worked around that by converting EXPECT_EQ(a, b) to EXPECT_TRUE(b == a).  Another possibility is to define the constants in the .cc file and merely declare them in the header, which works but is more verbose and less readable.

The original motiviation for this change was to find a way to eliminate some cases of passing anonymous-typed values as template arguments (which happens when you use a value from the enum in e.g. EXPECT_EQ()), which is technically illegal in C++03, though we don't warn about it.  Simply naming the enum would have done this, but this would have encouraged readers to actually use the enum name as a type, which for a bitfield is inappropriate for the reason given in the first paragraph.

BUG=92247
TEST=Compiles
Review URL: http://codereview.chromium.org/7756007

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@99088 0039d316-1c4b-4281-b951-d872f2087c98