Fix crash by preventing extensions from changing their IDs.
I can trigger the CHECK in 96466 by manually changing the 'key' field in the copy of an extension's manifest stored in the Preferences. With this change, the ExtensionService will stop loading the extension and instead print an error message if an extension's ID changes. This should not happen under normal conditions, so I'm not positive this is causing the crashes seen on real machines. 

BUG=96466
TEST=none


Review URL: http://codereview.chromium.org/7981002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@102939 0039d316-1c4b-4281-b951-d872f2087c98