Give each process its own bootstrap subset port as a subset of its inherited bootstrap port, and use the bootstrap subset port as the bootstrap port.
This completely eliminates leaks of on-demand Mach services advertised via the
bootstrap server.

This also reverts r34318 and r34534, "temporary" (21-month) hacks to mitigate
the leak. The temporary hacks were never completely effective against Breakpad
ports leaking from child processes.

DestructCrashReporter at process shutdown was deemed unnecessary and is being
removed. As this was the last caller to that function, the implementation is
removed as well.

This is addressed in Chrome rather than Breakpad to account for the potential
leak of rohitfork ports if the browser process crashes or is mercilessly
killed, because library code messing with the task's bootstrap port doesn't
strike me as kosher, and because the Mac Breakpad code is scheduled to be
replaced with something better that doesn't attempt to leak ports like a sieve
within a couple of months anyway.

BUG=28547
TEST=1. "launchctl bslist" should no longer show on-demand
        com.Breakpad.Inspector ports (in Breakpad-enabled builds with Breakpad
        on) or com.google.Chrome.rohitfork, com.google.Chrome.canary.rohitfork,
        or org.chromium.Chromium.rohitfork ports.
     2. "launchctl bstree" (as root) should reveal a bootstrap subset for the
        browser process as a child of the per-user/per-session bootstrap
        namespace containing the rohitfork port and browser's Breakpad port if
        crash reporting is on. There should also be a bootstrap subset for 
        each child process as a child of the browser's bootstrap subset. If
        crash reporting is on, each child process' bootstrap subset should 
        contain a Breakpad port.
     3. Breakpad reports should be generated on crashes. For example,
        about:crash and about:inducebrowsercrashforrealz should each cause a
        minidump to be written in
        ~/Library/Application Support/Google/Chrome/Crash Reports
        when Breakpad is enabled. This tests that the Breakpad ports are
        functioning properly.
     4. The browser process should be able to access child process data.
        Window:Task Manager should show valid values for the Memory, CPU, and
        Network columns for all child processes. This tests that the rohitfork
        port is functioning properly.
Review URL: http://codereview.chromium.org/8059041

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@103089 0039d316-1c4b-4281-b951-d872f2087c98