CrOS - Speculative fix for shutdown hang talking to zygote
ZygoteHost in the browser acquires a lock, sends fork requests to the zygote, then blocks on a read while waiting for a reply with the pid of the forked process.  In many error conditions, the zygote never sends a reply to the browser.  Now we send back an invalid pid (-1), which ZygoteHost already handles.

BUG=chromium-os:20825
TEST=Ensure you can open and close tabs normally.  After that, we'll have to watch crash reports.

Review URL: http://codereview.chromium.org/8062001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@103467 0039d316-1c4b-4281-b951-d872f2087c98