PPAPI SetFullscreen: Explicitely disallow state changes when the opposite one is already pending.
It is not clear what one should expect. Should the 2nd request override the 1st one? 
Should it be queued up and processed as if it was placed after the 1st one finished?
Should it be ignored? In reality, the 2nd transitions does not occur while still
messing up the state of 1st one (e.g. the window goes into fullscreen, but fullscreen
DidChangeView never fires). Even with the JS interface, the behavior appears to be
broken: http://www/~polina/fullscreen.html. Since the feature is not particularly useful,
it is cleanest to just deny it in Pepper.

BUG=41780
TEST=in CL
Review URL: http://codereview.chromium.org/8093013

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@103767 0039d316-1c4b-4281-b951-d872f2087c98