chromeos: Make Search key never close the NTP.
The Search key had complicated behavior: pressing it would
open the new tab page and focus its omnibox... unless the
NTP was already open, in which case the key would close it.
This was intended to provide an escape hatch for accidental
presses, but it prevents users from just mashing the key and
typing a URL whenever they want to load a page.

This change gets rid of the closing behavior -- if the NTP
is already open, we just focus the omnibox.

BUG=chromium-os:8742
TEST=added a simple unit test (which doesn't cover the focusing change); also did manual testing: opened NTP, unfocused omnibox, and checked that Search key re-focuses it

Review URL: http://codereview.chromium.org/8093012

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@103793 0039d316-1c4b-4281-b951-d872f2087c98