chromeos: Add halo to secure/VPN network badge.
This switches to a lock new image from kenmoore@ that
includes a white halo so it'll remain visible when using
dark themes.  The new image is a bit larger than the old
one, so I'm also updating the badging code to ensure that
badges don't get clipped.

BUG=chromium-os:21110
TEST=manual: made code always display vpn badge and checked that it looks fine against default and Johnny Cash themes


Review URL: http://codereview.chromium.org/8136020

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@104069 0039d316-1c4b-4281-b951-d872f2087c98