Remove unnecessary OSX specific IPC
Removing this IPC simplifies code paths across platforms and may make it less like to confuse error conditions in the renderer process.

When a tab is closed, the corresponding RWHV which responds to IPCs from the GPU process no longer exists. Many of these IPCs require an ACK IPC. As of r104092, the browser now sends an ACK when a RWHV no longer exists. This is temporary, once the renderer receives the browser signal for tear down, the GPU process will eventually stop sending IPCs for the defunct RWHV.

Previously, the IPCs were dropped and no ACK was sent. The solution was to tell the GPU process to tear down the resources associated with the defunt RWHV. This led to two tear down signals reaching the renderer (with possibly different semantics) -- one directly from the browser to the renderer and another from the browser to the GPU process to the render (shows up as a channel error).

BUG=none
TEST=open 2 tabs with http://www.webkit.org/blog-files/3d-transforms/poster-circle.html; close one; other still animates

Review URL: http://codereview.chromium.org/8146003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@104292 0039d316-1c4b-4281-b951-d872f2087c98