First pass at "async" signin flow.
Updated SyncSetupFlow to immediately step to the DONE/ENTER_PASSPHRASE state if
it can determine that a passphrase is/is not needed, rather than waiting for
a sync cycle to complete.

Updated ProfileSyncService to track the last auth error received even if there
are no encrypted types enabled (previously it would throw away that auth error
in that case, which made it impossible for the UI to determine if a passphrase
error would be exposed by enabling a data type).

BUG=19993, 95939
TEST=Run sync signin test cases.


Review URL: http://codereview.chromium.org/8165001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@104355 0039d316-1c4b-4281-b951-d872f2087c98