Allow whitelisted bundles to load in Chrome:
Default Folder X as of CFBundleVersion 4.4.3
MySpeed as of CFBundleVersion 1201
SIMBL (Albert Zeyer's fork) at any CFBundleVersion
Smart Scroll as of CFBundleVersion 3.9

This also removes the prohibition on loading SIMBL plug-ins. If a permitted
SIMBL is present, it is resposnbile for loading plug-ins and provides the
interface for plug-ins to override code.

BUG=95045
TEST=These products should now work in Chrome
Review URL: http://codereview.chromium.org/8226026

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@105251 0039d316-1c4b-4281-b951-d872f2087c98