Initialize the NetworkDelayListener to allow requests rather than deny them.
This is necessary for Chrome Frame, which doesn't load extensions.

We're relying on the Profile being created (and thus loading extensions) before the session is restored on startup. And as a result, we don't need to watch for NOTIFICATION_EXTENSIONS_READY anymore, nor test that.

TBR=aa

BUG=99450
TEST=NetworkDelayListener unit tests pass, and chrome_frame_net_tests don't time out
Review URL: http://codereview.chromium.org/8314005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@105687 0039d316-1c4b-4281-b951-d872f2087c98