Suppress bubble view for CONNECTION_FAILED error
If there was a connection failed error then we woud display an error bubble view and wrench menu item.

We're not supposed to display error messages for CONNECTION_FAILED errors so all the labels in the bubble view and wrench menu error item were blank.

The problem was that we would decide when to show the error UI based on the auth error state.

To fix this I simplified the code so that we show the error UI based directly on the presence of the error labels. This makes the code significantly simpler.

I also added unit tests to make sure that we're returning labels when we should not not returning them when we shouldn't.

BUG=100451
TEST=


Review URL: http://codereview.chromium.org/8308005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@105944 0039d316-1c4b-4281-b951-d872f2087c98