Prevent tooltips from appearing when the close button is gone.
When closing the find bar and downloads tray, the tooltip for the close
button will appear after the button itself is gone. Cocoa seems to get
confused when the view is animated out from under the mouse. Switch the
code to use a method that allows Chrome to decide if the tooltip should
be shown or not.

BUG=100850
TEST=Manual. Test that tooltip does not appear after closing the Find Bar,
infobars, or download tray. Test that hovering over the close button *does*
still show the tooltip in all these cases.


Review URL: http://codereview.chromium.org/8353037

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@106710 0039d316-1c4b-4281-b951-d872f2087c98