Calls to rph updating existing handlers are ignored.
This happens when a handler is registered, where there is already a handler
with the same origin and scheme already registered. These calls should go
through the the registry and update the registered handler, and now do.

This was broken recently however the unit tests did not catch the problem,
because the key logic was in browser.cc. That logic has been moved into the
registry and a test added.

BUG=99926
TEST=Unit test added.


Review URL: http://codereview.chromium.org/8371023

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@107052 0039d316-1c4b-4281-b951-d872f2087c98