This change implements certificate import from ONC file.
It will import any certificates in an ONC file into the certificate
database.  It does not yet associate the given GUID for the
certificate with the certificate added to the database.  That will
happen in an upcoming CL.

This also adds test cases for importing each type of certificate.

BUG=chromium-os:19404,chromium-os:19406
TEST=Ran new unit test cases.  Imported ONC file with a certificate
     and saw that it appeared in the certificate manager.


Review URL: http://codereview.chromium.org/8403007

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@108551 0039d316-1c4b-4281-b951-d872f2087c98