Gather history of capture and encode time determine next recoring delay
The current logic of chromoting is to capture as fast as possible with
20fps as a hard limit. This creates a heavy load on the slower systems with
only one processor.

I measured 100% CPU load running on these systems, resulting in very
jerky performance. In addition to speed up capture and encode we should
use system load and configurations as an information to determine recording
frequency.

In this patch I used the number of processors and history of capture and
encode time to determine the next recording delay. This essentially lower
recoding frequency when system is under load or is incapable to catch up
with the current 20fps frequency.

I have tested this patch on a Linux system limited to single core and
intentionally added extra 70ms of while(1) loop in capture and encode to
simulate slow processing. The results are as follows:

             CPU load       Latency
Before           100%         440ms
After            ~50%         360ms

So as a result both CPU load and latency improved.

BUG=100314
TEST=remoting_unittests and manually test with with heavily loaded system

Review URL: http://codereview.chromium.org/8342040

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@108871 0039d316-1c4b-4281-b951-d872f2087c98